+++ services/web/tmtar/events/controller.py
@@ -15,6 +15,8 @@ api = Namespace(
     decorators=[cross_origin()],
 )
 
+admin_location_field = 'admin_location_id'
+
 
 @api.route('/')
 class EventsResource(Resource):
@@ -28,16 +30,16 @@ class EventsResource(Resource):
         inject_claims=True,
     )
     def post(self, claims):
-        """Get all events for editing."""
+        """Create new event instance."""
         return SecureEventService.create(
             request.parsed_obj,
-            user_admin_location_id=claims['admin_location_id'],
+            user_admin_location_id=claims[admin_location_field],
         )
 
 
 @api.route('/all/<int:page>')
 class PaginatedEventsResource(Resource):
-    """Events resource."""
+    """Paginated Events resource."""
 
     @responds(schema=EventSchema(many=True), api=api)
     @access_restriction(
@@ -48,8 +50,60 @@ class PaginatedEventsResource(Resource):
     def get(self, page: int, claims):
         """Get all events for editing."""
         return SecureEventService.get_all(
-            user_admin_location_id=claims['admin_location_id'],
+            user_admin_location_id=claims[admin_location_field],
             page=page,
+        )
+
+
+@api.route('/all/search/<string:str_to_find>')
+@api.param('str_to_find', 'Part of Event title to search')
+class EventsSearchResource(Resource):
+    """Providing Events search."""
+
+    @api.doc(responses={
+        200: '{"status": "Match",\n "events": [Event Model object]}',
+        404: '{"status": "No match"}',
+    })
+    @access_restriction(
+        required_role=Role.admin,
+        api=api,
+        inject_claims=True,
+    )
+    def get(self, str_to_find: str, claims):
+        """Get matching events by part of title."""
+        events = SecureEventService.search_by_title(
+            str_to_find,
+            user_admin_location_id=claims[admin_location_field],
+        )
+        if events:
+            serialized_events = EventSchema().dump(events, many=True)
+            return jsonify(
+                {'status': 'Match', 'events': serialized_events},
+            )
+        return jsonify({'status': 'No match'}), 404
+
+
+@api.route('/all/count')
+class EventsCountResource(Resource):
+    """Events count resource."""
+
+    @api.doc(
+        responses={
+            200: "{'count': int}",
+        },
+    )
+    @access_restriction(
+        required_role=Role.admin,
+        api=api,
+        inject_claims=True,
+    )
+    def get(self, claims):
+        """Get count of all accessible events."""
+        events_count = SecureEventService.count_all_events(
+            user_admin_location_id=claims[admin_location_field],
+        )
+        return jsonify(
+            {'count': events_count},
         )
 
 
@@ -108,7 +162,7 @@ class EventsIdResource(Resource):
         return SecureEventService.update_by_id(
             event_id,
             changes,
-            user_admin_location_id=claims['admin_location_id'],
+            user_admin_location_id=claims[admin_location_field],
         )
 
     @api.doc(
@@ -125,7 +179,7 @@ class EventsIdResource(Resource):
         """Delete single Event by id."""
         deleted_id = SecureEventService.delete_by_id(
             event_id,
-            user_admin_location_id=claims['admin_location_id'],
+            user_admin_location_id=claims[admin_location_field],
         )
 
         return jsonify({'status': 'Success', 'id': deleted_id})